update anomaly test, rename class
